Combine model shim and model loader in models.py
